fix: move `HelpDialog` outside of `Popover` to ensure proper behavior
The `HelpDialog` component was previously nested inside the `Popover`, leading to unexpected behavior when attempting to close the dialog. By moving the `HelpDialog` component outside the `Popover`, the components no longer interfere with each other in the DOM tree, resolving the issue.
